refactor: 메세지 상수 관심사 분리
