+++ src/view/signup/AccountRegister/ProfileCreation.tsx
@@ -1,11 +1,12 @@
 import Button from "@/common/component/Button";
 import { FormController } from "@/shared/component/Form";
+import { useCheckOnServer } from "@/shared/hook/useCheckOnServer";
 import { handleOnChangeMode } from "@/shared/util/form";
 import { controllerStyle, formContainer } from "@/view/signup/index.css";
 import type { UseFormReturn } from "react-hook-form";
+import { defaultSignupMsg } from "./useSignupForm";
 
 type ProfileCreationProps = {
-  isActive: boolean;
   form: UseFormReturn<
     {
       password: string;
@@ -17,14 +18,25 @@ type ProfileCreationProps = {
     any,
     undefined
   >;
-  nicknameMsg: string;
 };
 
-const ProfileCreation = ({
-  isActive,
-  form,
-  nicknameMsg,
-}: ProfileCreationProps) => {
+const ProfileCreation = ({ form }: ProfileCreationProps) => {
+  const nickname = form.watch("nickname");
+
+  const { isNicknameLoading } = useCheckOnServer(form, nickname);
+  const { isValid, errors, dirtyFields } = form.formState;
+
+  const showNicknameMsg =
+    !(errors.nickname || isNicknameLoading) && dirtyFields.nickname;
+
+  const nicknameMsg = isNicknameLoading
+    ? defaultSignupMsg.nicknameLoading
+    : showNicknameMsg
+      ? defaultSignupMsg.validNickname
+      : errors.nickname?.message || defaultSignupMsg.nickname;
+
+  const isActive = isValid && !isNicknameLoading;
+
   return (
     <>
       <div className={formContainer}>
