+++ src/view/signup/AccountRegister/index.tsx
@@ -1,10 +1,7 @@
 "use client";
-import { signUpAction } from "@/app/api/auth/actions";
 import { Form } from "@/shared/component/Form";
-import type { signupSchema } from "@/view/signup/AccountRegister/schema";
 import { containerStyle } from "@/view/signup/index.css";
 import { useState } from "react";
-import type { z } from "zod";
 import Stepper from "../Stepper";
 import { SIGNUP_STEPS } from "../constant";
 import PasswordSetup from "./PasswordSetup";
@@ -12,29 +9,10 @@ import ProfileCreation from "./ProfileCreation";
 import useSignupForm from "./useSignupForm";
 
 const AccountRegister = ({ token }: { token: string }) => {
+  const { form, handleSubmit } = useSignupForm(token);
+
   const [step, setStep] = useState(SIGNUP_STEPS.PASSWORD_SETUP);
-  const { form, passwordError, passwordMsg, nicknameMsg, isActive } =
-    useSignupForm();
-
   const handleNextStep = () => setStep(SIGNUP_STEPS.PROFILE_CREATION);
-
-  const handleSubmit = async (values: z.infer<typeof signupSchema>) => {
-    const data = new FormData();
-
-    if (values.profile) {
-      data.append("profileImage", values.profile);
-    }
-
-    data.append(
-      "request",
-      JSON.stringify({
-        password: values.password,
-        nickname: values.nickname,
-      }),
-    );
-
-    await signUpAction(token, data);
-  };
 
   return (
     <div>
@@ -48,18 +26,9 @@ const AccountRegister = ({ token }: { token: string }) => {
           className={containerStyle}
         >
           {step === SIGNUP_STEPS.PASSWORD_SETUP ? (
-            <PasswordSetup
-              form={form}
-              passwordError={passwordError}
-              passwordMsg={passwordMsg}
-              onNextStep={handleNextStep}
-            />
+            <PasswordSetup form={form} onNextStep={handleNextStep} />
           ) : (
-            <ProfileCreation
-              isActive={isActive}
-              form={form}
-              nicknameMsg={nicknameMsg}
-            />
+            <ProfileCreation form={form} />
           )}
         </form>
       </Form>
